Refactoring for `DelegateCommand`, `AsyncCommand` and the `CommandExtensions` are finished.
