+++ src/MLib3.MVVM/Commands/DelegateCommand.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Input;
+
+namespace MLib3.MVVM
+{
+    public class DelegateCommand : DelegateCommand<object>
+    {
+        public DelegateCommand(Action<object?> execute, Func<object?, bool> canExecute) : base(execute, canExecute)
+        {
+        }
+    }
+}
