+++ src/MLib3.MVVM/Commands/ReactionApi/IPropertyChangedHandler.cs
@@ -1,0 +1,8 @@
+﻿using System.ComponentModel;
+
+namespace MLib3.MVVM;
+
+public interface IPropertyChangedHandler : IDisposable
+{
+    void Observe(INotifyPropertyChanged? source);
+}
