--- src/MLib3.MVVM/ICommandReactionSubPath.cs
@@ -1,7 +1,0 @@
-﻿namespace MLib3.MVVM
-{
-    public interface ICommandReactionSubPath : ICommandReactionPath
-    {
-        ICommandReactionSubPath And(string property);
-    }
-}
