+++ src/MLib3.MVVM/ViewModel.cs
@@ -5,19 +5,9 @@ namespace MLib3.MVVM;
 
 public class ViewModel : IViewModel
 {
-    public event PropertyChangedEventHandler PropertyChanged;
+    public event PropertyChangedEventHandler? PropertyChanged;
 
     protected virtual void OnPropertyChanged([CallerMemberName] string? property = null)
-    {
-        if (property == null) 
-            throw new ArgumentNullException(nameof(property));
-        var propertyInfo = GetType().GetProperty(property);
-        if (propertyInfo == null) 
-            throw new ArgumentException($"Property {property} not found in {GetType().Name}");
-        OnPropertyChanged(propertyInfo.GetValue(this), property);
-    }
-    
-    protected virtual void OnPropertyChanged(object? newValue, [CallerMemberName] string? property = null)
     {
         if (property == null) 
             throw new ArgumentNullException(nameof(property));
@@ -31,7 +21,7 @@ public class ViewModel : IViewModel
         var newValue = value;
         field = newValue;
         callback?.Invoke(oldValue, newValue);
-        OnPropertyChanged(newValue, property);
+        OnPropertyChanged(property);
     }
 
     protected virtual void Set<T>(Func<T> get, Action<T> set, T value, ValueChangedCallback<T>? callback = null, [CallerMemberName] string? property = null)
@@ -41,7 +31,7 @@ public class ViewModel : IViewModel
         var newValue = value;
         set(newValue);
         callback?.Invoke(oldValue, newValue);
-        OnPropertyChanged(newValue, property);
+        OnPropertyChanged(property);
     }
     
     protected virtual T Get<T>(Func<T> get) => get();
