+++ tests/MLib3.MVVM.UnitTests/Mocks/A.cs
@@ -1,0 +1,19 @@
+﻿using System.ComponentModel;
+
+namespace MLib3.MVVM.UnitTests.Mocks;
+
+public class A : INotifyPropertyChanged
+{
+    public event PropertyChangedEventHandler PropertyChanged;
+
+    private B _b = new B();
+    public B B
+    {
+        get => _b;
+        set
+        {
+            _b = value;
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(B)));
+        }
+    }
+}
